requirements.txt
flask
flask-cors
sympy
numpy
pytesseract
opencv-python
pillow
gunicorn